If an exception in Cygdb occurs, a stacktrace will be printed.
