+++ Cython/Debugger/libcython.py
@@ -11,7 +11,6 @@ except NameError:
 
 import sys
 import textwrap
-import traceback
 import functools
 import itertools
 import collections
@@ -68,19 +67,6 @@ _filesystemencoding = sys.getfilesystemencoding() or 'UTF-8'
 
 
 # decorators
-
-def dont_suppress_errors(function):
-    "*sigh*, readline"
-    @functools.wraps(function)
-    def wrapper(*args, **kwargs):
-        try:
-            return function(*args, **kwargs)
-        except Exception:
-            traceback.print_exc()
-            raise
-
-    return wrapper
-
 
 def default_selected_gdb_frame(err=True):
     def decorator(function):
@@ -695,7 +681,7 @@ class CyImport(CythonCommand):
     command_class = gdb.COMMAND_STATUS
     completer_class = gdb.COMPLETE_FILENAME
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     def invoke(self, args, from_tty):
         if isinstance(args, BYTES):
             args = args.decode(_filesystemencoding)
@@ -842,7 +828,7 @@ class CyBreak(CythonCommand):
             if func.pf_cname:
                 gdb.execute('break %s' % func.pf_cname)
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     def invoke(self, function_names, from_tty):
         if isinstance(function_names, BYTES):
             function_names = function_names.decode(_filesystemencoding)
@@ -861,7 +847,7 @@ class CyBreak(CythonCommand):
             else:
                 self._break_funcname(funcname)
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     def complete(self, text, word):
         # Filter init-module functions (breakpoints can be set using
         # modulename:linenumber).
@@ -952,7 +938,7 @@ class CyStep(CythonExecutionControlCommand, libpython.PythonStepperMixin):
     name = 'cy -step'
     stepinto = True
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     def invoke(self, args, from_tty):
         if self.is_python_function():
             self.python_step(self.stepinto)
@@ -982,7 +968,7 @@ class CyRun(CythonExecutionControlCommand):
 
     name = 'cy run'
 
-    invoke = dont_suppress_errors(CythonExecutionControlCommand.run)
+    invoke = libpython.dont_suppress_errors(CythonExecutionControlCommand.run)
 
 
 class CyCont(CythonExecutionControlCommand):
@@ -992,7 +978,7 @@ class CyCont(CythonExecutionControlCommand):
     """
 
     name = 'cy cont'
-    invoke = dont_suppress_errors(CythonExecutionControlCommand.cont)
+    invoke = libpython.dont_suppress_errors(CythonExecutionControlCommand.cont)
 
 
 class CyFinish(CythonExecutionControlCommand):
@@ -1001,7 +987,7 @@ class CyFinish(CythonExecutionControlCommand):
     """
     name = 'cy finish'
 
-    invoke = dont_suppress_errors(CythonExecutionControlCommand.finish)
+    invoke = libpython.dont_suppress_errors(CythonExecutionControlCommand.finish)
 
 
 class CyUp(CythonCommand):
@@ -1011,7 +997,7 @@ class CyUp(CythonCommand):
     name = 'cy up'
     _command = 'up'
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     def invoke(self, *args):
         try:
             gdb.execute(self._command, to_string=True)
@@ -1046,7 +1032,7 @@ class CySelect(CythonCommand):
 
     name = 'cy select'
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     def invoke(self, stackno, from_tty):
         try:
             stackno = int(stackno)
@@ -1073,7 +1059,7 @@ class CyBacktrace(CythonCommand):
     command_class = gdb.COMMAND_STACK
     completer_class = gdb.COMPLETE_NONE
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     @require_running_program
     def invoke(self, args, from_tty):
         # get the first frame
@@ -1107,7 +1093,7 @@ class CyList(CythonCommand):
     command_class = gdb.COMMAND_FILES
     completer_class = gdb.COMPLETE_NONE
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     # @dispatch_on_frame(c_command='list')
     def invoke(self, _, from_tty):
         sd, lineno = self.get_source_desc()
@@ -1124,7 +1110,7 @@ class CyPrint(CythonCommand):
     name = 'cy print'
     command_class = gdb.COMMAND_DATA
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     def invoke(self, name, from_tty):
         print("repr("+name+") = ", end="")
         self.cy.exec_.invoke("print(" + name + ")", from_tty)
@@ -1149,7 +1135,7 @@ class CyLocals(CythonCommand):
     command_class = gdb.COMMAND_STACK
     completer_class = gdb.COMPLETE_NONE
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     @dispatch_on_frame(c_command='info locals', python_command='py-locals')
     def invoke(self, args, from_tty):
         cython_function = self.get_cython_function()
@@ -1181,7 +1167,7 @@ class CyGlobals(CyLocals):
     command_class = gdb.COMMAND_STACK
     completer_class = gdb.COMPLETE_NONE
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     @dispatch_on_frame(c_command='info variables', python_command='py-globals')
     def invoke(self, args, from_tty):
         global_python_dict = self.get_cython_globals_dict()
@@ -1319,7 +1305,7 @@ class CyExec(CythonCommand, libpython.PyExec, EvaluateOrExecuteCodeMixin):
     command_class = gdb.COMMAND_STACK
     completer_class = gdb.COMPLETE_NONE
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     def invoke(self, expr, from_tty):
         expr, input_type = self.readcode(expr)
         executor = libpython.PythonCodeExecutor()
@@ -1342,7 +1328,7 @@ class CySet(CythonCommand):
     command_class = gdb.COMMAND_DATA
     completer_class = gdb.COMPLETE_NONE
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     @require_cython_frame
     def invoke(self, expr, from_tty):
         name_and_expr = expr.split('=', 1)
@@ -1366,7 +1352,7 @@ class CyCName(gdb.Function, CythonBase):
         print $cy_cname("module.function")
     """
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     @require_cython_frame
     @gdb_function_value_to_unicode
     def invoke(self, cyname, frame=None):
@@ -1398,7 +1384,7 @@ class CyCValue(CyCName):
     Get the value of a Cython variable.
     """
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     @require_cython_frame
     @gdb_function_value_to_unicode
     def invoke(self, cyname, frame=None):
@@ -1419,7 +1405,7 @@ class CyLine(gdb.Function, CythonBase):
     Get the current Cython line.
     """
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     @require_cython_frame
     def invoke(self):
         return self.get_cython_lineno()
@@ -1430,7 +1416,7 @@ class CyEval(gdb.Function, CythonBase, EvaluateOrExecuteCodeMixin):
     Evaluate Python code in the nearest Python or Cython frame and return
     """
 
-    @dont_suppress_errors
+    @libpython.dont_suppress_errors
     @gdb_function_value_to_unicode
     def invoke(self, python_expression):
         input_type = libpython.PythonCodeExecutor.Py_eval_input
